requirements.txt
blinker==1.8.2
Brotli==1.0.9
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.7
cx-Oracle==8.3.0
cycler==0.11.0
easyocr==1.7.1
filelock==3.13.1
Flask==3.0.3
fonttools==4.51.0
fsspec==2024.6.1
idna==3.7
imageio==2.34.2
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.4
lazy_loader==0.4
MarkupSafe==2.1.3
matplotlib==3.9.0
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
opencv-python-headless==4.10.0.84
oracledb==2.2.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
ply==3.11
pyclipper==1.3.0.post5
pycparser==2.22
pyparsing==3.0.9
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.9.0post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
scikit-image==0.24.0
scipy==1.14.0
setuptools==69.5.1
shapely==2.0.4
sip==6.7.12
six==1.16.0
sympy==1.12
tbb==2021.13.0
tifffile==2024.6.18
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tornado==6.4.1
typing_extensions==4.11.0
tzdata==2024.1
unicodedata2==15.1.0
urllib3==2.2.2
Werkzeug==3.0.3
wheel==0.43.0
win-inet-pton==1.1.0
